Added support for custom VTuber subscription Also moved the existing subscription command code here.
