+++ bot.py
@@ -11,9 +11,10 @@ from discord.ext import commands
 from ext.infoscraper import channelInfo
 from ext.cogs.subCycle import StreamCycle, streamcheck
 from ext.cogs.msCycle import msCycle, milestoneNotify
-from ext.share.botUtils import subPerms, chunks, creatorCheck
+from ext.share.botUtils import subPerms, creatorCheck
 from ext.share.dataGrab import getSubType, getwebhook
 from ext.share.prompts import botError, subCheck
+from ext.commands.subscribe import subCustom, subHolo
 
 with open("data/settings.yaml") as f:
     settings = yaml.load(f, Loader=yaml.SafeLoader)
@@ -77,125 +78,11 @@ async def subdef_error(ctx, error):
 
 @bot.command(aliases=['sub'])
 @commands.check(subPerms)
-async def subscribe(ctx):
-    listmsg = await ctx.send("Loading channels list...")
-
-    with open("data/channels.json", encoding="utf-8") as f:
-        channels = json.load(f)
-    csplit = []
-    for split in chunks(channels, 9):
-        csplit.append(split)
-    pagepos = 0
-    
-    while True:
-        picknum = 1
-        pickstr = ""
-        picklist = []
-        for split in csplit[pagepos]:
-            ytch = csplit[pagepos][split]
-            pickstr += f'{picknum}. {ytch["name"]}\n'
-            picklist.append(split)
-            picknum += 1
-        if pagepos == 0:
-            pickstr += f'\nA. Subscribe to all channels\nN. Go to next page\nX. Cancel'
-        elif pagepos == len(csplit) - 1:
-            pickstr += f'\nA. Subscribe to all channels\nB. Go to previous page\nX. Cancel'
-        else:
-            pickstr += f'\nA. Subscribe to all channels\nN. Go to next page\nB. Go to previous page\nX. Cancel'
-
-        listembed = discord.Embed(title="Subscribe to channel:", description=pickstr)
-        await listmsg.edit(embed=listembed)
-
-        def check(m):
-            return m.content.lower() in ['1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'n', 'b', 'x'] and m.author == ctx.author
-
-        while True:
-            try:
-                msg = await bot.wait_for('message', timeout=60.0, check=check)
-            except asyncio.TimeoutError:
-                await listmsg.delete()
-                await ctx.message.delete()
-                return
-            if msg.content in ['1', '2', '3', '4', '5', '6', '7', '8', '9']:
-                with open("data/servers.json") as f:
-                    servers = json.load(f)
-                await getwebhook(bot, servers, ctx.guild, ctx.channel)
-                with open("data/servers.json") as f:
-                    servers = json.load(f)
-                await msg.delete()
-                if "subDefault" not in servers[str(ctx.guild.id)][str(ctx.channel.id)]:
-                    uInput = await subCheck(ctx, bot, listmsg, 1, csplit[pagepos][picklist[int(msg.content) - 1]]["name"])
-                else:
-                    uInput = {
-                        "success": True,
-                        "subType": servers[str(ctx.guild.id)][str(ctx.channel.id)]["subDefault"]
-                    }
-                lastSubbed = False
-
-                if not uInput["success"]:
-                    await listmsg.delete()
-                    await ctx.message.delete()
-                    return
-
-                for subType in uInput["subType"]:
-                    if picklist[int(msg.content) - 1] not in servers[str(ctx.guild.id)][str(ctx.channel.id)][subType]:
-                        servers[str(ctx.guild.id)][str(ctx.channel.id)][subType].append(picklist[int(msg.content) - 1])
-                    else:
-                        if len(uInput["subType"]) > 1 and not lastSubbed:
-                            lastSubbed = True
-                        else:
-                            ytch = csplit[pagepos][picklist[int(msg.content) - 1]]
-                            await listmsg.edit(content=f'This channel is already subscribed to {ytch["name"]}.', embed=None)
-                            await ctx.message.delete()
-                            return
-                with open("data/servers.json", "w") as f:
-                    json.dump(servers, f, indent=4)
-                ytch = csplit[pagepos][picklist[int(msg.content) - 1]]
-                await listmsg.edit(content=f'This channel is now subscribed to: {ytch["name"]}.', embed=None)
-                await ctx.message.delete()
-                return
-            elif msg.content.lower() == 'a':
-                with open("data/servers.json") as f:
-                    servers = json.load(f)
-                await getwebhook(bot, servers, ctx.guild, ctx.channel)
-                with open("data/servers.json") as f:
-                    servers = json.load(f)
-                await msg.delete()
-                if "subDefault" not in servers[str(ctx.guild.id)][str(ctx.channel.id)]:
-                    uInput = await subCheck(ctx, bot, listmsg, 1, "Subscribing to all channels.")
-                else:
-                    uInput = {
-                        "success": True,
-                        "subType": servers[str(ctx.guild.id)][str(ctx.channel.id)]["subDefault"]
-                    }
-                if not uInput["success"]:
-                    await listmsg.delete()
-                    await ctx.message.delete()
-                    return
-                for subType in uInput["subType"]:
-                    for ytch in channels:
-                        if ytch not in servers[str(ctx.guild.id)][str(ctx.channel.id)][subType]:
-                            servers[str(ctx.guild.id)][str(ctx.channel.id)][subType].append(ytch)
-                with open("data/servers.json", "w") as f:
-                    json.dump(servers, f, indent=4)
-                await listmsg.edit(content=f'This channel is now subscribed to all Hololive YouTube channels.', embed=None)
-                await ctx.message.delete()
-                return
-            elif msg.content.lower() == 'n' and pagepos < len(csplit) - 1:
-                await msg.delete()
-                pagepos += 1
-                break
-            elif msg.content.lower() == 'b' and pagepos > 0:
-                await msg.delete()
-                pagepos -= 1
-                break
-            elif msg.content.lower() == 'x':
-                await msg.delete()
-                await listmsg.delete()
-                await ctx.message.delete()
-                return
-            else:
-                await msg.delete()
+async def subscribe(ctx, *, customUser = None):
+    if customUser is None:
+        await subHolo(ctx, bot)
+    else:
+        await subCustom(ctx, bot, customUser)
 
 @subscribe.error
 async def sub_error(ctx, error):
